+++ crates/net/common/src/lib.rs
@@ -9,4 +9,4 @@
 
 pub mod ban_list;
 pub mod bandwidth_meter;
-pub mod metered_channel;
+pub mod metered_sender;
